+++ e2e/create-certificate.test.ts
@@ -4,7 +4,6 @@ import {
   AcmeAuthorization,
   AcmeClient,
   AcmeOrder,
-  Dns01Challenge,
   DnsUtils,
 } from "../src/mod.ts";
 import { resolveDns } from "../src/resolveDns.deno.ts";
@@ -41,7 +40,7 @@ it("can talk to ACME server and successfully create an account, order then all t
 
   const dns01Challenge = authorization.findDns01Challenge();
   expectToBeDefined(dns01Challenge);
-  expect(dns01Challenge instanceof Dns01Challenge).toBe(true);
+  expect(dns01Challenge.type).toBe("dns-01");
   console.log("✅ Order > Authorization(s) > dns-01 challenge (found!)");
 
   const expectedRecord = await dns01Challenge.getDnsRecordAnswer();